src/stress/requirements.txt
locust==2.37.11
python-dotenv==1.0.0
